Refactored AttachmentsSelect
